ref/well_dashboard/requirements.txt
# FastAPI Backend
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiosqlite==0.19.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Streamlit Frontend
streamlit==1.32.0

# Data Processing
pandas==2.2.0
numpy==1.26.3
openpyxl==3.1.2
xlsxwriter==3.1.9

# Visualization
plotly==5.18.0

# HTTP Client for Streamlit to API
httpx==0.26.0

# Utilities
python-dateutil==2.8.2
